use OpenAI API for issue classification
