+++ .github/issue-classifier/README.md
@@ -1,8 +1,8 @@
 # Issue classifier
 
-The issue classifier uses GitHub Copilot CLI to apply the repository's existing
-labels to new and updated issues. It is disabled until the repository variable
-`ISSUE_CLASSIFIER_ENABLED` is set to `true`.
+The issue classifier uses the OpenAI Responses API to apply the repository's
+existing labels to new and updated issues. It is disabled until the repository
+variable `ISSUE_CLASSIFIER_ENABLED` is set to `true`.
 
 ## Behavior
 
@@ -12,17 +12,17 @@ labels to new and updated issues. It is disabled until the repository variable
   `COLLABORATOR`. Pull request comments are ignored.
 - Model context contains the current issue followed by trusted comments,
   newest first. Content and comment counts are capped by `config.json`.
-- Copilot uses automatic model selection with its minimum supported 30-credit
-  session cap. A custom agent exposes no tools, and the workflow disables
-  built-in MCP servers and repository instructions before sending untrusted
-  issue text. A normal classification is a single model response; the cap is a
-  backstop, not an expected cost.
+- The classifier uses `gpt-5.6-luna` with reasoning disabled, strict structured
+  output, no tools, and response storage disabled. Untrusted issue text is kept
+  separate from the fixed classifier instructions. A normal classification is
+  one Responses API request.
 - A decision must meet the configured confidence threshold before it changes a
   label. Low-confidence decisions leave the current state unchanged.
 - The workflow may remove only labels most recently applied by
   `github-actions[bot]`. Human additions are preserved, and a human removal is
   a sticky override.
-- The workflow never comments on an issue, closes it, or creates labels.
+- The workflow never gives the model tools, comments on an issue, closes it, or
+  creates labels.
 
 ## Managed labels
 
@@ -47,19 +47,20 @@ The workflow's manual dispatch supports two modes:
 
 Manual dispatches bypass the automatic per-issue quota. Set `dry_run` to false
 only when intentionally reconciling the selected issue. The evaluation makes
-40 Copilot requests and consumes organization-billed AI credits.
+40 Responses API requests and consumes OpenAI API usage.
 
 ## Rollout
 
-1. An organization owner enables Copilot CLI and selects **Allow use of
-   Copilot CLI billed to the organization**. No long-lived secret is needed;
-   the workflow uses its scoped `GITHUB_TOKEN`.
-2. Merge this workflow while `ISSUE_CLASSIFIER_ENABLED` is absent or false.
-3. Run the unit-test workflow, the historical evaluation, and a manual dry run.
-4. Set `ISSUE_CLASSIFIER_ENABLED` to `true` in repository Actions variables.
-5. After verifying a live maintainer-created issue, enable public issue
+1. Create a dedicated OpenAI service account in a project with an appropriate
+   budget and rate limits.
+2. Add its key as an organization Actions secret named `OPENAI_API_KEY`, with
+   access limited to `vala-lang/vala-language-server`. The pull-request test job
+   does not receive this secret.
+3. Merge this workflow while `ISSUE_CLASSIFIER_ENABLED` is absent or false.
+4. Run the unit-test workflow, the historical evaluation, and a manual dry run.
+5. Set `ISSUE_CLASSIFIER_ENABLED` to `true` in repository Actions variables.
+6. After verifying a live maintainer-created issue, enable public issue
    creation in the repository's Issues settings.
 
-Configure an organization cost center and budget before enabling automatic
-runs. Set `ISSUE_CLASSIFIER_ENABLED` to `false` as the kill switch if accuracy,
-credit use, or label churn is unacceptable.
+Set `ISSUE_CLASSIFIER_ENABLED` to `false` as the kill switch if accuracy, API
+usage, or label churn is unacceptable.
